+++ src/ArduinoJson/Strings/Adapters/RamString.hpp
@@ -26,8 +26,8 @@ class RamString {
   static constexpr size_t sizeMask = size_t(-1);
 #endif
 
-  RamString(const char* str, size_t sz, bool linked = false)
-      : str_(str), size_(sz & sizeMask), linked_(linked) {
+  RamString(const char* str, size_t sz, bool isStatic = false)
+      : str_(str), size_(sz & sizeMask), static_(isStatic) {
     ARDUINOJSON_ASSERT(size_ == sz);
   }
 
@@ -49,8 +49,8 @@ class RamString {
     return str_;
   }
 
-  bool isLinked() const {
-    return linked_;
+  bool isStatic() const {
+    return static_;
   }
 
  protected:
@@ -59,10 +59,10 @@ class RamString {
 #if ARDUINOJSON_SIZEOF_POINTER <= 2
   // Use a bitfield only on 8-bit microcontrollers
   size_t size_ : sizeof(size_t) * 8 - 1;
-  bool linked_ : 1;
+  bool static_ : 1;
 #else
   size_t size_;
-  bool linked_;
+  bool static_;
 #endif
 };